refactor: move fetchFilteredTasks to task-dashboard-helpers and update imports
